requirements.txt
Flask==0.10.1
Flask-Login==0.3.2
Flask-OAuthlib==0.9.2
Flask-RESTful==0.3.5
oauthlib==1.1.1
pydash==2.4.2
Werkzeug==0.11.9
factory_boy==2.7.0
factoryboy-gaendb==0.1.5
Unidecode==0.4.19
GoogleAppEngineCloudStorageClient==1.9.22.1
piplapis-python==5.0.8
